Basic Authentication and RBAC

+++ backend/controllers/goalController.js
@@ -10,6 +10,9 @@ export const createGoal = async (req, res, next) => {
     if (!goalData.title) {
       return res.status(400).json({ message: 'Title is required' });
     }
+
+    goalData.user = req.user.id;
+
     const goal = await goalService.createGoal(goalData);
     res.status(201).json(goal);
   } catch (error) {
@@ -39,10 +42,18 @@ export const getGoalById = async (req, res, next) => {
     if (!id) {
       return res.status(400).json({ message: 'Goal ID is required' });
     }
+
     const goal = await goalService.getGoalById(id);
     if (!goal) {
       return res.status(404).json({ message: 'Goal not found' });
     }
+
+    if (goal.user.toString() !== req.user.id) {
+      return res
+        .status(403)
+        .json({ message: 'Forbidden: You do not have accesss to this goal' });
+    }
+
     res.status(200).json(goal);
   } catch (error) {
     next(error);
@@ -63,10 +74,18 @@ export const updateGoal = async (req, res, next) => {
     if (!updateData || Object.keys(updateData).length === 0) {
       return res.status(400).json({ message: 'Update data is required' });
     }
-    const updatedGoal = await goalService.updateGoal(id, updateData);
-    if (!updatedGoal) {
+
+    const goal = await goalService.getGoalById(id);
+    if (!goal) {
       return res.status(404).json({ message: 'Goal not found' });
     }
+    if (goal.user.toString() !== req.user.id) {
+      return res.status(403).json({
+        message: 'Forbidden: You cannot update a goal that is not yours',
+      });
+    }
+
+    const updatedGoal = await goalService.updateGoal(id, updateData);
     res.status(200).json(updatedGoal);
   } catch (error) {
     next(error);
@@ -83,6 +102,17 @@ export const deleteGoal = async (req, res, next) => {
     if (!id) {
       return res.status(400).json({ message: 'Goal ID is required' });
     }
+
+    const goal = await goalService.getGoalById(id);
+    if (!goal) {
+      return res.status(404).json({ message: 'Goal not found' });
+    }
+    if (goal.user.toString() !== req.user.id) {
+      return res.status(403).json({
+        message: 'Forbidden: You cannot delete a goal that is not yours',
+      });
+    }
+
     await goalService.deleteGoal(id);
     // 204 No Content indicates successful deletion with no response body.
     res.status(204).send();
